Extract logic for creating an org into hook
